+++ server/src/Classes/Routes/contract.php
@@ -17,6 +17,12 @@ class Contract implements ControllerProviderInterface
 
 		/* search */
         $controllers->get('/', function (Application $app, Request $request) {
+        	
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'contract_view')){
+        		return $app->json(array("totalCount"=>0, "contract"=>array()));
+        	}
+        	
             $sort = '';
     		if ($request->get('sort')){
     			$sort = json_decode($request->get('sort'), true);
@@ -29,7 +35,7 @@ class Contract implements ControllerProviderInterface
     		if ($request->get('search')){
     			$search = json_decode($request->get('search'), true);
     		}
-            list($totalCount, $result) = $app['business.contract']->getAll($request->get('page'),$request->get('start'),$request->get('limit'),$sort,$filter,$request->get('query'),$search);
+            list($totalCount, $result) = $app['business.contract']->getAll($app, $request->get('page'),$request->get('start'),$request->get('limit'),$sort,$filter,$request->get('query'),$search);
 
             return $app->json(array("totalCount"=>$totalCount, "contract"=>$result));
         });
@@ -53,6 +59,11 @@ class Contract implements ControllerProviderInterface
         });
 
         $controllers->get('/{id}', function(Application $app, $id, Request $request) {
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'contract_view')){
+        		return $app->json(array("success"=>false, "totalCount"=>0, "contract"=>array()));
+        	}
+        	
             $contract = $app['business.contract']->getById($id);
 
 			$contract['client'] = $app['business.client']->getById($contract['client_id']);
@@ -66,6 +77,11 @@ class Contract implements ControllerProviderInterface
 
 		// create
 		$controllers->post('/new', function(Application $app, Request $request) {
+			// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'contract_create')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+        	
             $params = json_decode($request->getContent(),true);
             $error = $app['business.contract']->validate($app, $params);
             if (@count($error) > 0){
@@ -78,6 +94,11 @@ class Contract implements ControllerProviderInterface
 
 		// update
         $controllers->put('/{id}', function(Application $app, $id, Request $request) {
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'contract_edit')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+        	
         	$params = json_decode($request->getContent(),true);
             $error = $app['business.contract']->validate($app, $params);
             if (@count($error) > 0){
@@ -89,6 +110,11 @@ class Contract implements ControllerProviderInterface
         });
 
         $controllers->delete('/delete/{id}', function(Application $app, $id, Request $request) {
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'contract_delete')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+        	
             //$params = json_decode($request->getContent(),true);
             $app['business.contract']->deleteContract($app, $id);
             $app['business.advertisement']->deleteByContractId($id);
