+++ server/src/Classes/Routes/territory.php
@@ -16,34 +16,49 @@ class Territory implements ControllerProviderInterface
         $controllers = $app['controllers_factory'];
 
         $controllers->get('/', function (Application $app, Request $request) {
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'territory_view')){
+        		return $app->json(array("totalCount"=>0, "territory"=>array()));
+        	}
+        	
             $sort = '';
-        		if ($request->get('sort')){
-        			$sort = json_decode($request->get('sort'), true);
-        		}
-        		$filter = array();
-        		if ($request->get('filter')){
-        			$filter = json_decode($request->get('filter'), true);
-        		}
-        		$search = array();
-        		if ($request->get('search')){
-        			$search = json_decode($request->get('search'), true);
-        		}
-            list($totalCount, $result) = $app['business.territory']->getAll($request->get('page'),$request->get('start'),$request->get('limit'),$sort,$filter,$request->get('query'),$search);
+    		if ($request->get('sort')){
+    			$sort = json_decode($request->get('sort'), true);
+    		}
+    		$filter = array();
+    		if ($request->get('filter')){
+    			$filter = json_decode($request->get('filter'), true);
+    		}
+    		$search = array();
+    		if ($request->get('search')){
+    			$search = json_decode($request->get('search'), true);
+    		}
+            list($totalCount, $result) = $app['business.territory']->getAll($app, $request->get('page'),$request->get('start'),$request->get('limit'),$sort,$filter,$request->get('query'),$search);
             return $app->json(array("totalCount"=>$totalCount, "territory"=>$result));
         });
 				
-				$controllers->get('/{id}', function (Application $app, $id, Request $request) {
+		$controllers->get('/{id}', function (Application $app, $id, Request $request) {
+			// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'territory_view')){
+        		return $app->json(array("totalCount"=>0, "territory"=>array()));
+        	}
+			
             $result = $app['business.territory']->getById($id);
             if ($result['id']){
-	            $result['state'] = $app['business.state']->getById($result['state_id']);
-	            $result['manager'] =  $app['business.user']->getById($result['manager_id']);
-	          }
+				$result['state'] = $app['business.state']->getById($result['state_id']);
+				$result['manager'] =  $app['business.user']->getById($result['manager_id']);
+			}
             return $app->json(array("success"=>true,"totalCount"=>(isset($result['id'])?1:0), "territory"=>$result));
         });
 		    
         $controllers->post('/new', function(Application $app, Request $request) {
-        		$params = json_decode($request->getContent(),true);
-        		$error = $app['business.territory']->validate($app, $params);
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'territory_create')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+        	
+    		$params = json_decode($request->getContent(),true);
+    		$error = $app['business.territory']->validate($app, $params);
             if (@count($error) > 0){
             	return $app->json(array("success"=>false,"error"=>$error));
             } else {
@@ -53,6 +68,11 @@ class Territory implements ControllerProviderInterface
         });
 
         $controllers->put('/{id}', function(Application $app, $id, Request $request) {
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'territory_edit')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+        	
             $params = json_decode($request->getContent(),true);
             $error = $app['business.territory']->validate($app, $params);
             if (@count($error) > 0){
@@ -64,6 +84,11 @@ class Territory implements ControllerProviderInterface
         });
 
         $controllers->delete('/delete/{id}', function(Application $app, $id, Request $request) {
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'territory_delete')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+        	
             $result = $app['business.territory']->deleteTerritory($id);
             return $app->json(array("success"=>true,"territory"=>$result));
         });
