+++ server/src/Classes/Routes/user.php
@@ -16,6 +16,12 @@ class User implements ControllerProviderInterface
 
 		/* search */
         $controllers->get('/', function (Application $app, Request $request) {
+        	
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'user_view')){
+        		return $app->json(array("totalCount"=>0, "user"=>array()));
+        	}
+        	
     	    $sort = '';
     		if ($request->get('sort')){
     			$sort = json_decode($request->get('sort'), true);
@@ -28,7 +34,7 @@ class User implements ControllerProviderInterface
     		if ($request->get('search')){
     			$search = json_decode($request->get('search'), true);
     		}
-            list($totalCount, $user_array) = $app['business.user']->getAll($request->get('page'),$request->get('start'),$request->get('limit'),$sort,$filter,$request->get('query'),$search);
+            list($totalCount, $user_array) = $app['business.user']->getAll($app, $request->get('page'),$request->get('start'),$request->get('limit'),$sort,$filter,$request->get('query'),$search);
             //$totalCount = $app['business.user']->getTotalCount($request->get('filter'));
 
             array_walk($user_array,function($user,$key) use (&$user_array, &$app){
@@ -41,14 +47,25 @@ class User implements ControllerProviderInterface
 
 		/* get */
 	    $controllers->get('/{id}', function(Application $app, $id, Request $request) {
+	    	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'user_view')){
+        		return $app->json(array("success" => false, "totalCount"=>0, "user"=>array()));
+        	}
+        	
 	        $user_array = $app['business.user']->getById($id);
 	        $user_array['roles'] = $app['business.permissionrole']->getByUser($user_array['id']);
 	        $user_array['manager'] = $app['business.user']->getById($user_array['manager_user_id']);
+	        $user_array['regional'] = $app['business.user']->getById($user_array['regional_user_id']);
 	        return $app->json(array("success"=>true,"totalCount"=>($user_array['id']?1:0),"user"=>$user_array));
 	    });
 
 		/* update */
 		$controllers->put('/{id}', function(Application $app, $id, Request $request) {
+			// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'user_edit')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+			
             $params = json_decode($request->getContent(),true);
             $error = $app['business.user']->validate($app, $params);
             if (@count($error) > 0){
@@ -73,12 +90,22 @@ class User implements ControllerProviderInterface
 
         /* delete */
 		$controllers->delete('/delete/{id}', function(Application $app, $id, Request $request) {
+			// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'user_delete')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+			
             $user_array = $app['business.user']->deleteUser($id);
             return $app->json(array("success"=>true,"user"=>$user_array));
         });
 
 		/* create */
         $controllers->post('/new', function(Application $app, Request $request) {
+        	// make sure we have the right permission
+        	if (!$app['business.user']->hasPermission($app, 'user_create')){
+        		return $app->json(array("success"=>false,"error"=>array('Invalid permission')));
+        	}
+        	
             $params = json_decode($request->getContent(),true);
             $error = $app['business.user']->validate($app, $params);
             if (@count($error) > 0){
